==> 5. Separando os eventos.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from pandas import Timedelta, Timestamp
from collections import namedtuple

from IPython.core.debugger import set_trace

df_estacao = pd.read_csv("DATA/dados_estacoes_5min.csv",
        parse_dates=["DATE"],
        index_col=["DATE"])

df_radar = pd.read_csv("DATA/dados_radar_semHorarioVerao.csv",
                      parse_dates=['DATE'],
                      index_col=["DATE"])

df_estacao.sort_index(inplace=True)
df_radar.sort_index(inplace=True)

df_est = df_estacao[df_estacao > 0].dropna(how="all",axis=0).dropna(how="all",axis=1)
indices = df_est.index
eventos = []

evento = []
last_indice = indices[0]
evento.append(last_indice)
for indice in indices[1:]:
    evento.append(indice)
    # Eventos com um gap maior que 10 min são eventos distintos
    if indice - last_indice > Timedelta("10min"):
        # eventos maiores que 10 min
        if len(evento) > 1:
            eventos.append(evento)
        evento = []
    last_indice = indice
print(len(eventos))
    

# Evento aleátorio para tests
evento = eventos[3]
evento

df_est.loc[evento].dropna(axis=1, how="all")

df_evento = df_est.loc[evento].dropna(axis=1, how="all")
evento_estacao = df_evento.MB_PRO1.dropna()
evento_estacao

evento_estacao.sum()

evento_estacao.index

def plot_event(radar,station,name):
    plt.figure()
    plt.title(name)
    station.plot(label="STATION")
    radar.plot(label="RADAR")
    plt.legend()
    
    file_name = "IMAGENS/eventos/" + str(name)+"jpg"
    plt.savefig(file_name)

É necessario ainda separar os eventos por estação, como é possivel ver acima os eventos começam em horarios distintos por estaçoes.


def test_min_pluvi(serie, indice):
    test_indice = np.arange(indice - Timedelta("6h"),indice,Timedelta("10min"))
    test_min_pluvi = serie.loc[test_indice].sum()
    
    return test_min_pluvi > 1

from IPython.core.debugger import set_trace
df = pd.DataFrame(columns=["start", "station","duration","Pluviometro","RADAR"])
stations = df_est.columns
for station in stations:
    df_station = df_est[station]
    df_station_nonull = df_station[df_station > 0].dropna()
    indices = df_station_nonull.index
    dict_evento = {}
    evento = []
    last_indice = indices[0]
    for indice in indices[1:]:
        evento.append(last_indice)
        #set_trace()
        # Eventos com um gap maior que 6 horas são eventos distintos
        # E com menos de 1 mm de chuva
        
        
        
        diff_ = indice - evento[0]
                 
        if  indice - last_indice > Timedelta("6h") or (diff_ >= Timedelta("6h") and test_min_pluvi(df_station,indice)):
            total_event = df_station.loc[evento].sum()
              
            dict_evento["start"] = evento[0]
            dict_evento["station"] = station
            dict_evento["duration"] = (evento[-1] - evento[0]) + Timedelta("10min")
            dict_evento["Pluviometro"] = total_event
            try:
                dict_evento["RADAR"] = df_radar[station].loc[evento].sum()
                name = station + str(evento[0])
                #plot_event(df_radar[station].loc[evento],df_station.loc[evento],name)
            except KeyError:
                dict_evento["RADAR"] = np.nan

            df = df.append(dict_evento, ignore_index=True)
            evento = []
        last_indice = indice

df.head()

#df = df[df.Pluviometro > 2.5]

df.reset_index(drop=True,inplace=True)

df["ERROR"] = df.Pluviometro - df.RADAR

df.describe()

df[df.Pluviometro == df.Pluviometro.max()]

df.drop(labels=df.Pluviometro.idxmax(),axis=0,inplace=True)

df.groupby("station").describe()["ERROR"]

df.to_csv("DATA/eventos6h.csv",index=False)
